dom/media/webaudio/test/test_convolverNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the ConvolverNode interface</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var context = new AudioContext();
  var conv = new ConvolverNode(context);

  is(conv.channelCount, 2, "Convolver node has 2 input channels by default");
  is(conv.channelCountMode, "clamped-max", "Correct channelCountMode for the Convolver node");
  is(conv.channelInterpretation, "speakers", "Correct channelCountInterpretation for the Convolver node");

  is(conv.buffer, null, "Default buffer value");
  conv.buffer = context.createBuffer(2, 1024, 22050);
  is(conv.normalize, true, "Default normalize value");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
